@@ -1,7 +1,10 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule, FormControl, FormGroup, Validators} from '@angular/forms';
+import { CommonModule } from '@angular/common';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { AppRoutingModule } from './app-routing.module';
+import { ToastrModule } from 'ngx-toastr';
 import { AppComponent } from './app.component';
 import { OrdersComponent } from './orders/orders.component';
 import { ListComponent } from './orders/list/list.component';
@@ -12,8 +15,7 @@ import { DisplayValidationDirective } from './display-validation.directive';
 import { UpdateItemComponent } from './orders/update-item/update-item.component';
 import { ItemDetailsComponent } from './orders/item-details/item-details.component';
 import { OrderService} from './orders/order.service';
-import { OrderItemService} from './orders/orderItem.service';
-import { Http, Headers,HttpModule} from '@angular/http';
+import { HttpClientModule } from '@angular/common/http';
 
 @NgModule({
   declarations: [
@@ -26,18 +28,18 @@ import { Http, Headers,HttpModule} from '@angular/http';
     DisplayValidationDirective,
     UpdateItemComponent,
     ItemDetailsComponent,
-
   ],
   imports: [
     BrowserModule,
     AppRoutingModule,
     FormsModule,
     ReactiveFormsModule,
-    HttpModule
+    HttpClientModule,
+    CommonModule,
+    BrowserAnimationsModule, // required animations module
+    ToastrModule.forRoot(),
      ],
-  providers: [OrderService, 
-              HttpModule,
-              OrderItemService],
+  providers: [OrderService],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
